requirements.txt
# Quiz Bot Pro - Python Dependencies
# Python 3.11+

# ============================================================
# CORE - Telegram Bot Framework
# ============================================================
aiogram==3.13.1
aiohttp>=3.9.0

# ============================================================
# DATABASE
# ============================================================
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.14.0
greenlet>=3.0.0

# ============================================================
# REDIS
# ============================================================
redis[hiredis]==5.2.1

# ============================================================
# VALIDATION & SETTINGS
# ============================================================
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# ============================================================
# LOGGING
# ============================================================
structlog==24.4.0

# ============================================================
# AUDIO (TTS)
# ============================================================
gTTS==2.5.4

# Optional Azure TTS (uncomment if needed)
# azure-cognitiveservices-speech==1.35.0

# ============================================================
# SCHEDULING
# ============================================================
APScheduler==3.10.4

# ============================================================
# UTILITIES
# ============================================================
python-dateutil>=2.8.0
tenacity>=8.2.0

# ============================================================
# DEVELOPMENT
# ============================================================
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
black==24.10.0
ruff==0.8.3
mypy==1.13.0

# ============================================================
# TYPE STUBS
# ============================================================
types-redis>=4.6.0
types-python-dateutil>=2.8.0
